// src/services/openapi/api.ts
export interface CategoryEntity {
    /**
     * 
     * @type {string}
     * @memberof CategoryEntity
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof CategoryEntity
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof CategoryEntity
     */
    'description': string;
    /**
     * 
     * @type {boolean}
     * @memberof CategoryEntity
     */
    'isActive': boolean;
}
// ...
export interface CreateProductDto {
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'productId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'productTypeId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'categoryId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'subCategoryId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'manufacturerId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'returnExchangePolicy': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'skuid': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'productName': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'productDescription': string;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'offerprice': number;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'actualprice': number;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isActive': boolean;
    /**
     * 
     * @type {CreateSareeDetailsDto}
     * @memberof CreateProductDto
     */
    'productDetails': CreateSareeDetailsDto;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'cgst': number;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'sgst': number;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'maxQuantityPerCart': number;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'promoId'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isTrending': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isBestSeller': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isExclusive': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isShippable': boolean;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'collectionId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'colorId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'occassionId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'styleId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'printId'?: string;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'maxAllowedReturnDays': number;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'maxAllowedCancellationDays': number;
}
// ...
export interface CreateSareeDetailsDto {
    /**
     * 
     * @type {string}
     * @memberof CreateSareeDetailsDto
     */
    'fabricDetailsId': string;
    /**
     * 
     * @type {number}
     * @memberof CreateSareeDetailsDto
     */
    'length': number;
    /**
     * 
     * @type {number}
     * @memberof CreateSareeDetailsDto
     */
    'width': number;
    /**
     * 
     * @type {boolean}
     * @memberof CreateSareeDetailsDto
     */
    'isBlousePieceIncluded': boolean;
    /**
     * 
     * @type {string}
     * @memberof CreateSareeDetailsDto
     */
    'blouseDescription': string;
}
// ...
export interface FabricDetailsEntity {
    /**
     * 
     * @type {string}
     * @memberof FabricDetailsEntity
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof FabricDetailsEntity
     */
    'fabricName': string;
    /**
     * 
     * @type {string}
     * @memberof FabricDetailsEntity
     */
    'fabricDescription': string;
    /**
     * 
     * @type {string}
     * @memberof FabricDetailsEntity
     */
    'washCare': string;
}
/**
 * 
 * @export
 * @interface ManufacturerEntity
 */
export interface ManufacturerEntity {
    /**
     * 
     * @type {string}
     * @memberof ManufacturerEntity
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ManufacturerEntity
     */
    'origin': string;
    /**
     * 
     * @type {string}
     * @memberof ManufacturerEntity
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof ManufacturerEntity
     */
    'address': string;
}
// ...
export interface SubcategoryEntity {
    /**
     * 
     * @type {string}
     * @memberof SubcategoryEntity
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof SubcategoryEntity
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof SubcategoryEntity
     */
    'description': string;
    /**
     * 
     * @type {boolean}
     * @memberof SubcategoryEntity
     */
    'isActive': boolean;
    /**
     * 
     * @type {CategoryEntity}
     * @memberof SubcategoryEntity
     */
    'category': CategoryEntity;
}